refactor: move views to a new folder for views
