+++ frontend/src/main.tsx
@@ -1,6 +1,6 @@
 import ReactDOM from "react-dom/client";
 import { BrowserRouter } from "react-router";
-import { Root } from "./components/Root.tsx";
+import { Root } from "./views/Root.tsx";
 
 const root = document.getElementById("root");
 if (!root) throw new Error("Failed to find the root element");
