+++ frontend/src/views/App.tsx
@@ -1,5 +1,5 @@
-import Calendar from "./Calendar";
 import { StyledEngineProvider } from "@mui/styled-engine-sc";
+import Calendar from "../components/Calendar";
 
 function App() {
   return (
